refactor: extract api to barrel export #109
