+++ src/api/authClient/authClient.test.ts
@@ -1,0 +1,7 @@
+import { describe, expect, it } from 'vitest';
+
+describe('authClient - signup', () => {
+  it('should give success for first login', () => {
+    expect(false).toBe(true);
+  });
+});
