+++ src/api/cardClient/cardClient.test.ts
@@ -1,0 +1,7 @@
+import { describe, expect, it } from 'vitest';
+
+describe('authClient - signup', () => {
+  it('should get cards', () => {
+    expect(false).toBe(true);
+  });
+});
