+++ src/pages/SignIn/index.tsx
@@ -1,5 +1,5 @@
-import { Button, Checkbox, Input } from '../../Components';
-import { signInAPI } from '../../api/authClient';
+import { Button, Checkbox, Input } from '@/Components';
+import { signInAPI } from '@/api';
 import {
   useEmailSave,
   useInput,
@@ -7,7 +7,7 @@ import {
   useLogin,
 } from '@/hooks';
 import { useNavigate } from 'react-router-dom';
-import { ROUTE_PATHS } from '../../constant/config';
+import { ROUTE_PATHS } from '@/constant/config';
 import { useState } from 'react';
 import {
   ButtonWrapper,
@@ -18,7 +18,7 @@ import {
   Title,
 } from './SignIn.style';
 import { useMutation } from '@tanstack/react-query';
-import { checkEmail } from '../../utils';
+import { checkEmail } from '@/utils';
 
 /**
  * @todo 화면의 로직과 로그인과 관련된 비즈니스 로직이 강하게 결합되어 있습니다. 결합도를 나추도록 합니다.
